+++ ui/src/components/Dialogs/FileUploadDialog.js
@@ -35,7 +35,6 @@ export default class FileUploadDialog extends Component {
       uploadProgress: 10,
       loading: false
     };
-
   }
 
   handleChange(event) {
@@ -49,6 +48,16 @@ export default class FileUploadDialog extends Component {
     if (name === 'file-dialog') {
       console.log(event.target.files[0]);
       this.file = event.target.files[0];
+      if (this.file.name) {
+        const extension = this.file.name.split('.').pop().toUpperCase();
+
+        for (const resource of this.props.uploadOperation.resources) {
+          if (resource.extensions.map(st => st.toUpperCase()).indexOf(extension) !== -1) {
+            this.setState({file_type: resource.kind});
+          }
+        }
+      }
+
       this.setState({
         file_path: this.file,
         file_name: this.file ? this.file.name : null,
@@ -93,7 +102,7 @@ export default class FileUploadDialog extends Component {
         self.setState({loading: false});
       }
       if (this.props.onUpload) {
-        this.props.onUpload(data.node)
+        this.props.onUpload(data.node);
       }
     }).catch((error) => {
       console.error('error', error);
@@ -109,7 +118,7 @@ export default class FileUploadDialog extends Component {
           }
         });
       } else {
-        self.showAlert('Failed to upload file', 'failed');
+        self.props.showAlert('Failed to upload file. Did you choose the right file type?', 'failed');
       }
       self.setState({loading: false});
     });