+++ ui/src/components/LogInRedirect/index.js
@@ -34,7 +34,7 @@ export default class LogInRedirect extends Component {
     // Loading
 
     const self = this;
-    let loading = true;
+    const loading = true;
     let sleepPeriod = 1000;
     const sleepMaxPeriod = 10000;
     const sleepStep = 1000;
@@ -67,6 +67,9 @@ export default class LogInRedirect extends Component {
         cookie.save('user', response.data.user, { path: '/' });
         cookie.save('settings', response.data.settings, { path: '/' });
       };
+    } else if (specialUser === SPECIAL_USERS.REDIRECT) {
+      window.location = '/workflows/';
+      return;
     }
 
     const errorHandler = (error) => {